# Improves performance and reduce allocations in DefaultEndpointDataSource