Improves performance and reduce allocations in **DefaultEndpointDataSource**, and improves test coverage.

**Before:**

|                 Method |     Mean |     Error |    StdDev |         Op/s |  Gen 0 | Allocated |
|----------------------- |---------:|----------:|----------:|-------------:|-------:|----------:|
|     Constructor_Params | 63.06 ns | 0.2402 ns | 0.2247 ns | 15,859,022.6 | 0.0018 |     160 B |
| Constructor_Enumerable | 56.63 ns | 0.7778 ns | 0.7275 ns | 17,658,629.3 | 0.0017 |     160 B |

**After:**

|                 Method |     Mean |     Error |    StdDev |         Op/s |  Gen 0 | Allocated |
|----------------------- |---------:|----------:|----------:|-------------:|-------:|----------:|
|     Constructor_Params | 56.25 ns | 0.1883 ns | 0.1762 ns | 17,777,345.5 | 0.0014 |     120 B |
| Constructor_Enumerable | 51.05 ns | 0.2256 ns | 0.2110 ns | 19,590,400.7 | 0.0018 |     160 B |

Benchmark is available [here](https://gist.github.com/drieseng/8313f46e039a26a6b5655685dd6fa8c5).